requirements.txt
flask==3.0.3
flask-cors==4.0.1
streamlit==1.37.1
PyPDF2==3.0.1
sentence-transformers==3.0.1
pinecone-client==5.0.1
tiktoken==0.7.0
requests==2.32.3
python-dotenv==1.0.1
